Put Assignment files in their own folder & combine AssignmentForm tests
