+++ ui/src/Assignments/AssignmentCard/AssignmentCard.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
+@import '../../Styles/Colors';
 
 .personContainer {
   display: flex;
