+++ ui/src/Assignments/AssignmentCard/AssignmentCard.test.tsx
@@ -18,19 +18,19 @@
 import {fireEvent, screen, waitFor} from '@testing-library/react';
 import React from 'react';
 import AssignmentCard from './AssignmentCard';
-import {renderWithRecoil} from '../Utils/TestUtils';
-import TestData from '../Utils/TestData';
-import PeopleClient from '../Services/Api/PeopleClient';
-import {ViewingDateState} from '../State/ViewingDateState';
-import {IsReadOnlyState} from '../State/IsReadOnlyState';
-import {IsDraggingState} from '../State/IsDraggingState';
-import AssignmentClient from '../Services/Api/AssignmentClient';
+import {renderWithRecoil} from 'Utils/TestUtils';
+import TestData from 'Utils/TestData';
+import PeopleClient from 'Services/Api/PeopleClient';
+import {ViewingDateState} from 'State/ViewingDateState';
+import {IsReadOnlyState} from 'State/IsReadOnlyState';
+import {IsDraggingState} from 'State/IsDraggingState';
+import AssignmentClient from 'Services/Api/AssignmentClient';
 import moment from 'moment';
-import ProductClient from '../Services/Api/ProductClient';
-import {CurrentSpaceState} from '../State/CurrentSpaceState';
-import {RoleTag} from '../Types/Tag';
-import {Color} from '../Types/Color';
-import {Assignment} from '../Types/Assignment';
+import ProductClient from 'Services/Api/ProductClient';
+import {CurrentSpaceState} from 'State/CurrentSpaceState';
+import {RoleTag} from 'Types/Tag';
+import {Color} from 'Types/Color';
+import {Assignment} from 'Types/Assignment';
 
 jest.mock('Services/Api/ProductClient');
 jest.mock('Services/Api/AssignmentClient');
