+++ ui/src/Assignments/AssignmentCardList/AssignmentCardList.tsx
@@ -16,22 +16,22 @@
  */
 
 import React, {RefObject, useCallback, useRef, useState} from 'react';
-import AssignmentCard from 'Assignments/AssignmentCard';
+import AssignmentCard from 'Assignments/AssignmentCard/AssignmentCard';
 import {isUnassignedProduct} from 'Products/ProductService';
 import {
     getLocalStorageFiltersByType,
     personTagsFilterKey,
     roleTagsFilterKey,
-} from '../SubHeader/SortingAndFiltering/FilterLibraries';
+} from 'SubHeader/SortingAndFiltering/FilterLibraries';
 import {isPersonMatchingSelectedFilters} from 'People/PersonService';
-import useOnStorageChange from '../Hooks/useOnStorageChange/useOnStorageChange';
+import useOnStorageChange from 'Hooks/useOnStorageChange/useOnStorageChange';
 import {Draggable, Droppable} from 'react-beautiful-dnd';
 import {useRecoilValue} from 'recoil';
-import {IsReadOnlyState} from '../State/IsReadOnlyState';
-import {Product} from '../Types/Product';
-import {Assignment} from '../Types/Assignment';
+import {IsReadOnlyState} from 'State/IsReadOnlyState';
+import {Product} from 'Types/Product';
+import {Assignment} from 'Types/Assignment';
 
-import '../Products/Product.scss';
+import 'Products/Product.scss';
 
 interface Props {
     product: Product;
