+++ ui/src/Assignments/AssignmentExistingWarning/AssignmentExistsWarning.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
+@import '../../Styles/Colors';
 
 .warningText {
   font-size: 0.875rem;
