+++ ui/src/Assignments/PersonAndRoleInfo/PersonAndRoleInfo.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
+@import '../../Styles/Colors';
 
 .personNameAndRoleContainer {
   width: 142px;
