+++ ui/src/Assignments/UnassignedDrawer/UnassignedDrawer.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
+@import '../../Styles/Colors';
 
 .unassignedHeader {
   position: relative;
