+++ ui/src/Common/DrawerContainer/DrawerContainer.tsx
@@ -16,8 +16,8 @@
  */
 
 import React from 'react';
-import '../../Assignments/PersonDrawer.scss';
-import '../../Assignments/UnassignedDrawer.scss';
+import 'Assignments/PersonDrawer.scss';
+import 'Assignments/UnassignedDrawer/UnassignedDrawer.scss';
 import './DrawerContainer.scss';
 
 interface DrawerContainerProps {
