+++ ui/src/PeopleMover/PeopleMover.tsx
@@ -22,7 +22,7 @@ import Branding from '../Common/Branding/Branding';
 import SubHeader from '../SubHeader/SubHeader';
 import {useParams} from 'react-router-dom';
 import ReassignedDrawer from './ReassignedDrawer/ReassignedDrawer';
-import UnassignedDrawer from '../Assignments/UnassignedDrawer';
+import UnassignedDrawer from 'Assignments/UnassignedDrawer/UnassignedDrawer';
 import ArchivedProductsDrawer from '../Products/ArchivedProductsDrawer';
 import Counter from '../Common/Counter/Counter';
 import ArchivedPersonDrawer from '../People/ArchivedPersonDrawer';
