+++ ui/src/Products/ProductCard.tsx
@@ -19,23 +19,23 @@ import React, {useState} from 'react';
 import EditMenu, {EditMenuOption} from '../Common/EditMenu/EditMenu';
 import ProductClient from '../Services/Api/ProductClient';
 import {isUnassignedProduct} from './ProductService';
-import AssignmentCardList from '../Assignments/AssignmentCardList';
+import AssignmentCardList from 'Assignments/AssignmentCardList/AssignmentCardList';
 
 import moment from 'moment';
-import {createDataTestId} from '../Utils/ReactUtils';
-import {ProductPlaceholderPair} from '../Assignments/CreateAssignmentRequest';
-import AssignmentClient from '../Services/Api/AssignmentClient';
+import {createDataTestId} from 'Utils/ReactUtils';
+import {ProductPlaceholderPair} from 'Types/CreateAssignmentRequest';
+import AssignmentClient from 'Services/Api/AssignmentClient';
 import ConfirmationModal, {ConfirmationModalProps} from 'Modal/ConfirmationModal/ConfirmationModal';
 import {JSX} from '@babel/types';
-import {getAssignments} from '../People/PersonService';
+import {getAssignments} from 'People/PersonService';
 import {useRecoilValue, useSetRecoilState} from 'recoil';
-import {ViewingDateState} from '../State/ViewingDateState';
-import {IsReadOnlyState} from '../State/IsReadOnlyState';
-import {CurrentSpaceState} from '../State/CurrentSpaceState';
+import {ViewingDateState} from 'State/ViewingDateState';
+import {IsReadOnlyState} from 'State/IsReadOnlyState';
+import {CurrentSpaceState} from 'State/CurrentSpaceState';
 import useFetchProducts from 'Hooks/useFetchProducts/useFetchProducts';
-import {ModalContentsState} from '../State/ModalContentsState';
+import {ModalContentsState} from 'State/ModalContentsState';
 import ProductForm from './ProductForm';
-import AssignmentForm from '../Assignments/AssignmentForm';
+import AssignmentForm from 'Assignments/AssignmentForm/AssignmentForm';
 import {Product} from 'Types/Product';
 import {Person} from 'Types/Person';
